Amend normalization factors for each DOS pieces
Change the definition of the normalization factor for each DOS piece using the density of states at Eminindex (the lower energy bound of each window) instead of the global energy minimum E0.  This makes the normalization consistent within each energy window and affects the amount of "vertical shift" of the ln(g(E)), but it does not affect the correctness of the results as the ratio between the entries remains the same.